fix(lipsync): initialise sapi_lipsync members and widen bytes_to_milli
- Both constructors set only m_pPhnEstimator. close() frees m_pWaveFmt
  when it is non-null, and isDone() returns m_bDone, but loadAudio()
  is the only place that set either. When no SAPI recognizer is
  installed, initializeObjects() fails, loadAudio() never runs, and the
  destructor freed an indeterminate pointer. Both constructors now set
  m_pWaveFmt to null and m_bDone to false, and close() nulls the
  pointer after the free so a second close() is safe.
- bytes_to_milli multiplied a 32-bit byte count by 1000 in 32 bits,
  which overflows above about 4.29 MB (under 100 seconds of 44.1 kHz
  audio) and gave a wrong millisecond value. The multiply is now 64-bit.

No automated test: the class needs a SAPI recognizer and an audio
stream, and bytes_to_milli reads a protected member set by loadAudio.

Closes #70

+++ SCICompanionLib/Src/LipSync/sapi_lipsync.h
@@ -175,7 +175,9 @@ public:
 	 **/
     long bytes_to_milli(DWORD dwBytes)
     {
-        return (UINT)((dwBytes * 1000 )/ m_pWaveFmt->nAvgBytesPerSec); 
+        // Multiply in 64 bits: dwBytes * 1000 overflows 32 bits above about
+        // 4.29 MB, which is under 100 seconds of 44.1 kHz audio (#70).
+        return (long)(((ULONGLONG)dwBytes * 1000) / m_pWaveFmt->nAvgBytesPerSec);
     }
 
 protected:
